Edit and delete fixed
