+++ src/index.js
@@ -53,7 +53,7 @@ app.use((req, res, next) => {
 app.use(require('./routes'));
 app.use(require('./routes/authentication'));
 app.use('/links', require('./routes/index'));
-app.use('/links', require('./routes/links'));
+app.use('/',require('./routes/drogueria'))
