+++ src/views/links/add.hbs
@@ -4,7 +4,7 @@
 
             <div class="card">
                 <div class="card-body">
-                    <form action="/links/add" method="POST">
+                    <form action="/addProduct" method="POST">
                         <div class="form-group">
                             <input type="text" name="nombre" class="form-control" placeholder="nombre producto"
                                 autofocus>
