+++ src/views/links/edit.hbs
@@ -4,7 +4,7 @@
 
             <div class="card">
                 <div class="card-body">
-                    <form action="/links/edit/{{producto.id}}" method="POST">
+                    <form action="/edit/{{producto.id}}" method="POST">
                         <div class="form-group">
                             <input type="text" name="nombre" value="{{producto.nombre}}" class="form-control" placeholder="nombre producto"
                                 autofocus>
