+++ src/views/links/list.hbs
@@ -10,8 +10,8 @@
                         </h3>
                     </a>
                     <p class="m-2">{{descripcion}}</p>  
-                    <a href="/links/delete/{{id}}" class="btn btn-danger">Borrar Prodcuto</a>
-                    <a href="/links/edit/{{id}}" class="btn btn-secondary">EDITAR</a>
+                    <a href="/delete/{{id}}" class="btn btn-danger">Borrar Producto</a>
+                    <a href="/edit/{{id}}" class="btn btn-secondary">EDITAR</a>
 
                 </div>
             </div>
@@ -19,8 +19,8 @@
         {{else}}
             <div class="col-md-4 mx-auto">
                 <div class="card card-body text-center">
-                    <p>no hay productos aun</p>
-                    <a href="/links/add">Agrega uno</a>
+                    <p>No hay productos aún</p>
+                    <a href="/addProduct">Agregar uno</a>
 
                 </div>
             </div>
