+++ src/views/partials/nav-bar.hbs
@@ -29,7 +29,7 @@
         </li>
 
         <li class="nav-item mx-2 ms-auto">
-          <a class="nav-link" href="/productos" align="center">Productos</a>
+          <a class="nav-link" href="/producto" align="center">Productos</a>
         </li>
 
 
@@ -50,7 +50,7 @@
           </a>
           {{#each drogueria}}
           <ul class="dropdown-menu" aria-labelledby="navbarDropdownMenuLink">
-            <li><a class="dropdown-item" href="#">{{nombre}}</a></li>
+            <li><a class="dropdown-item" href="/droguerias">{{nombre}}</a></li>
           </ul>
           {{/each}}
         </li>
