--- src/views/partials/nav.hbs
@@ -1,51 +1,0 @@
-<nav class="navbar navbar-expand-lg navbar-light bg-light">
-
-  <div class="container ">
-
-    <div>
-      <a class="navbar-brand" href="/index">
-        <img src="/img/logo.svg" class="width-logo" alt="Sanar">Drogueria Sanar</a>
-    </div>
-
-    <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNavDropdown"
-      aria-controls="navbarNavDropdown" aria-expanded="false" aria-label="Toggle navigation">
-      <span class="navbar-toggler-icon"></span>
-    </button>
-    <div class="collapse navbar-collapse" id="navbarNavDropdown">
-      <ul class="navbar-nav d-flex">
-
-        {{#if user}}
-        <li class="nav-item mx-2">
-          <a class="nav-link" href="/list">Inventario</a>
-        </li>
-        <li class="nav-item mx-2">
-          <a class="nav-link" href="/add">Añadir Producto</a>
-        </li>
-        <li class="nav-item mx-2 ms-auto">
-          <a class="nav-link btn btn-success text-white" href="/logout">Cerrar sesion</a>
-        </li>
-
-        {{else}}
-
-        <li class="nav-item mx-2 ms-auto">
-          <a class="nav-link" href="#" align="center">Servicios</a>
-        </li>
-
-        <li class="nav-item mx-2 ms-auto">
-          <a class="nav-link" href="/productos" align="center">Productos</a>
-        </li>
-
-
-        <li class="nav-item mx-2 ms-auto">
-          <a class="nav-link" href="/signup">Registro</a>
-        </li>
-
-        <li class="nav-item mx-2 ms-auto">
-          <a class="nav-link" href="/signin" align="center">Iniciar sesion</a>
-        </li>
-
-        {{/if}}
-      </ul>
-    </div>
-  </div>
-</nav>
